down to 9 errors. NEVER DO THIS AGAIN!

+++ Engine/Game.cpp
@@ -43,16 +43,16 @@ void Game::UpdateModel()
 
 	// TODO : handle logic components.
 
-	float frameTime = m_clock.GetCounterS();
+	double frameTime = m_clock.GetCounterS();
 	m_clock.StartCounter();
 	GameWorldObject & activeGameWorldObject = m_camera;
-	float speed = 100.f*frameTime;
+	double speed = 100.f*frameTime;
 
 	Command * command = m_userInputManager.HandleInput();
 	if (command)
 	{
 		activeGameWorldObject.MoveBy({ 0,1 });
-		command->execute(activeGameWorldObject, speed);
+		command->execute(activeGameWorldObject, (float)speed);
 	}
 
 /*
